# docs/swagger.yaml
basePath: /api
definitions:
...
  ds.LoadCreateRequest:
    properties:
      load_category:
        type: string
      load_description:
        type: string
      load_title:
        type: string
      normative:
        type: number
      reliability_coefficient:
        type: number
    required:
    - load_category
    - load_description
    - load_title
    - normative
    - reliability_coefficient
    type: object
  ds.LoadDTO:
    properties:
      id:
        type: integer
      load_category:
        type: string
      load_description:
        type: string
      load_image:
        type: string
      load_title:
        type: string
      normative:
        type: number
      reliability_coefficient:
        type: number
      status:
        type: boolean
    type: object
...
  ds.LoadUpdateRequest:
    properties:
      load_category:
        type: string
      load_description:
        type: string
      load_title:
        type: string
      normative:
        type: number
      reliability_coefficient:
        type: number
    type: object
...
  ds.PaginatedResponse:
    properties:
      items: {}
      total:
        type: integer
    type: object
  ds.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API-сервер для управления нагрузками и сессиями расчета в системе.
  title: API для системы расчета нагрузок
  version: "1.0"
paths:
...
  /load-sessions/draft/loads/{load_id}:
    post:
      description: Добавляет нагрузку в черновик сессии текущего пользователя. Если
        черновика нет, создает новый.
      parameters:
      - description: ID нагрузки
        in: path
        name: load_id
        required: true
        type: integer
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ds.SuccessResponse'
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Добавить нагрузку в черновик (авторизованный пользователь)
      tags:
      - load-sessions
  /loads:
    get:
      description: Возвращает отфильтрованный список всех нагрузок. Доступно всем.
      parameters:
      - description: Поиск по названию
        in: query
        name: search
        type: string
      - description: Фильтр по категории
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.PaginatedResponse'
      summary: Получить список нагрузок (все)
      tags:
      - loads
    post:
      consumes:
      - application/json
      description: Создает новую нагрузку в системе. Требует прав модератора.
      parameters:
      - description: Данные нагрузки
        in: body
        name: load
        required: true
        schema:
          $ref: '#/definitions/ds.LoadCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ds.LoadDTO'
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Требуются права модератора
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Создать новую нагрузку (модератор)
      tags:
      - loads
  /loads/{id}:
    delete:
      description: Удаляет нагрузку из системы. Требует прав модератора.
      parameters:
      - description: ID нагрузки
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Требуются права модератора
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить нагрузку (модератор)
      tags:
      - loads
    get:
      description: Возвращает информацию о нагрузке по её ID. Доступно всем.
      parameters:
      - description: ID нагрузки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.LoadDTO'
        "404":
          description: Нагрузка не найдена
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить нагрузку по ID (все)
      tags:
      - loads
    put:
      consumes:
      - application/json
      description: Обновляет данные нагрузки. Требует прав модератора.
      parameters:
      - description: ID нагрузки
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: load
        required: true
        schema:
          $ref: '#/definitions/ds.LoadUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.LoadDTO'
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Требуются права модератора
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Обновить нагрузку (модератор)
      tags:
      - loads
  /loads/{id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Загружает изображение для нагрузки. Требует прав модератора.
      parameters:
      - description: ID нагрузки
        in: path
        name: id
        required: true
        type: integer
      - description: Изображение
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Требуются права модератора
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Загрузить изображение для нагрузки (модератор)
      tags:
      - loads
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"